refactor gRPC methods
